+++ lib/xue_hua_pinyin.dart
@@ -1,4 +1,3 @@
 library;
 
-export 'src/rust/api/pinyin_api.dart';
 export 'src/xue_hua_pinyin.dart';